# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "plot api",
    "contact": {}
  },
  "host": "localhost:4000",
  "basePath": "/",
  "securityDefinitions": {},
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/schools": {
      "post": {
        "summary": "create school",
        "tags": ["School"],
        "operationId": "createschool",
        "deprecated": false,
        "produces": ["application/json; charset=utf-8"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createschoolrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/createschool"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Created School"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "28"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"1c-IlOINX4DsY/GM7Cy8gmMODWW2ew\""
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              },
              "Date": {
                "type": "string",
                "default": "Tue, 28 Sep 2021 14:25:04 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/DuplicateError1"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "title": "Duplicate Error",
                "message": "School with this name or prefix exist"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "77"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"4d-i9q6eOUkHZd9KpHB2Yn302Na+AA\""
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              },
              "Date": {
                "type": "string",
                "default": "Tue, 28 Sep 2021 15:31:33 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Schools",
        "tags": ["School"],
        "operationId": "Schools",
        "deprecated": false,
        "produces": ["application/json; charset=utf-8"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "default": "",
              "items": {
                "$ref": "#/definitions/School"
              }
            },
            "examples": {
              "application/json; charset=utf-8": [
                {
                  "_id": "5f9183cad035174e94992476",
                  "name": "Brilliant Imapct School",
                  "schoolPrefix": "BIIA",
                  "address": "Angwan Tomato, Gauraka, Tafa L.G.A, Niger State ",
                  "phoneNumber": "+2348034055074",
                  "email": "dandynamicx@gmail.com",
                  "schoolId": "BIIA385589",
                  "__v": 1,
                  "section": "2020/2021",
                  "term": "First",
                  "faceBookBio": "https://facebook.com/",
                  "images": [],
                  "instagramBio": "https://www.instagram.com",
                  "notice": [],
                  "openedDays": 0,
                  "previousSchoolSection": ["2021/2022"],
                  "schoolImageAsBlob": " ",
                  "twitterBio": "https://twitter.com",
                  "youtubeBio": "https://youtube.com/"
                },
                {
                  "_id": "615325c063cda80d59c473e4",
                  "section": "2021/2022",
                  "term": "First",
                  "openedDays": 0,
                  "notice": [],
                  "previousSchoolSection": [],
                  "images": [],
                  "faceBookBio": "https://facebook.com/",
                  "twitterBio": "https://twitter.com",
                  "youtubeBio": "https://youtube.com/",
                  "instagramBio": "https://www.instagram.com",
                  "schoolImageAsBlob": " ",
                  "name": "Test",
                  "schoolPrefix": "XAx",
                  "address": "Angwan Tomato, Gauraka, Tafa L.G.A, Niger State",
                  "phoneNumber": "+2348034055074",
                  "email": "email",
                  "schoolId": "XAx704524",
                  "__v": 0
                }
              ]
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"400-CjxtzVyRvZ0yUOcga9gX4iljnTE\""
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              },
              "Content-Encoding": {
                "type": "string",
                "default": "gzip"
              },
              "Date": {
                "type": "string",
                "default": "Tue, 28 Sep 2021 14:25:12 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              },
              "Transfer-Encoding": {
                "type": "string",
                "default": "chunked"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete School",
        "tags": ["School"],
        "operationId": "DeleteSchool",
        "deprecated": false,
        "produces": ["application/json; charset=utf-8"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/DeleteSchoolRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DeleteSchool"
            },
            "examples": {
              "application/json; charset=utf-8": {
                "message": "Deleted School With Name Test"
              }
            },
            "headers": {
              "X-Powered-By": {
                "type": "string",
                "default": "Express"
              },
              "Access-Control-Allow-Origin": {
                "type": "string",
                "default": "*"
              },
              "Content-Length": {
                "type": "string",
                "default": "43"
              },
              "ETag": {
                "type": "string",
                "default": "W/\"2b-0LAbazv4CWpb8gBTGRfnjCeEj14\""
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              },
              "Date": {
                "type": "string",
                "default": "Tue, 28 Sep 2021 15:34:36 GMT"
              },
              "Connection": {
                "type": "string",
                "default": "keep-alive"
              },
              "Keep-Alive": {
                "type": "string",
                "default": "timeout=5"
              }
            }
          }
        }
      }
    },
...
    "/arm/5f8c7ee1b9776e05f105a6db": {
      "get": {
        "summary": "get schools",
        "tags": ["Arm"],
        "operationId": "getschools",
        "deprecated": false,
        "produces": ["application/json"],
        "consumes": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/result/addNotice": {
      "put": {
        "summary": "Add Notice",
        "tags": ["Result"],
        "operationId": "AddNotice",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AddNoticeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/result/removeNotice": {
      "delete": {
        "summary": "Delete Notice",
        "tags": ["Result"],
        "operationId": "DeleteNotice",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/DeleteNoticeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
    "createschoolrequest": {
      "title": "createschoolrequest",
      "example": {
        "name": "Test",
        "prefix": "XAx",
        "address": "Angwan Tomato, Gauraka, Tafa L.G.A, Niger State",
        "phoneNumber": "+2348034055074",
        "email": "email"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "prefix": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      },
      "required": ["name", "prefix", "address", "phoneNumber", "email"]
    },
    "createschool": {
      "title": "createschool",
      "example": {
        "message": "Created School"
      },
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": ["message"]
    },
    "DuplicateError1": {
      "title": "DuplicateError1",
      "example": {
        "title": "Duplicate Error",
        "message": "School with this name or prefix exist"
      },
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      },
      "required": ["title", "message"]
    },
    "School": {
      "title": "School",
      "example": {
        "_id": "5f9183cad035174e94992476",
        "name": "Brilliant Imapct School",
        "schoolPrefix": "BIIA",
        "address": "Angwan Tomato, Gauraka, Tafa L.G.A, Niger State ",
        "phoneNumber": "+2348034055074",
        "email": "dandynamicx@gmail.com",
        "schoolId": "BIIA385589",
        "__v": 1,
        "section": "2020/2021",
        "term": "First",
        "faceBookBio": "https://facebook.com/",
        "images": [],
        "instagramBio": "https://www.instagram.com",
        "notice": [],
        "openedDays": 0,
        "previousSchoolSection": ["2021/2022"],
        "schoolImageAsBlob": " ",
        "twitterBio": "https://twitter.com",
        "youtubeBio": "https://youtube.com/"
      },
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "schoolPrefix": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "schoolId": {
          "type": "string"
        },
        "__v": {
          "type": "integer",
          "format": "int32"
        },
        "section": {
          "type": "string"
        },
        "term": {
          "type": "string"
        },
        "faceBookBio": {
          "type": "string"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "instagramBio": {
          "type": "string"
        },
        "notice": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "openedDays": {
          "type": "integer",
          "format": "int32"
        },
        "previousSchoolSection": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "schoolImageAsBlob": {
          "type": "string"
        },
        "twitterBio": {
          "type": "string"
        },
        "youtubeBio": {
          "type": "string"
        }
      },
      "required": [
        "_id",
        "name",
        "schoolPrefix",
        "address",
        "phoneNumber",
        "email",
        "schoolId",
        "__v",
        "section",
        "term",
        "faceBookBio",
        "images",
        "instagramBio",
        "notice",
        "openedDays",
        "previousSchoolSection",
        "schoolImageAsBlob",
        "twitterBio",
        "youtubeBio"
      ]
    },
...
    "DeleteSchoolRequest": {
      "title": "DeleteSchoolRequest",
      "example": {
        "school": "615437fdbe92a442dca122ff"
      },
      "type": "object",
      "properties": {
        "school": {
          "type": "string"
        }
      },
      "required": ["school"]
    },
    "DeleteSchool": {
      "title": "DeleteSchool",
      "example": {
        "message": "Deleted School With Name Test"
      },
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": ["message"]
    },
...
    "AddNoticeRequest": {
      "title": "AddNoticeRequest",
      "example": {
        "message": "Just a test hshsxhshxhs",
        "school": "5f8c7ee1b9776e05f105a6db"
      },
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "school": {
          "type": "string"
        }
      },
      "required": ["message", "school"]
    },
    "DeleteNoticeRequest": {
      "title": "DeleteNoticeRequest",
      "example": {
        "message": "Just a test",
        "school": "5f8c7ee1b9776e05f105a6db"
      },
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "school": {
          "type": "string"
        }
      },
      "required": ["message", "school"]
    },
...
  },
...
}